optimazed contract
Hello David great project both on the frontend and contract. However i will love to point out some obvious observations. Starting from the adminAddress which should be a constructor accessing the any address and checking if it's same address with the admin. Next it's the admin address value it self. This address need to be hidden instead of hardcoding it in the contract. Next I added events to be emitted when a specific function is executed. This provide better logs out put of function calls. Event like newDrug() to emit on a new drug name and lenght on a success adding of drug. The drugBought() is like the opposite of newDrug() that emit the drug current owner, index and buyer address. The drugOutofStock() idea from aditya26sg work as an alert for empty stock. transferDrugOwnership() and transferStoreOwner() respectfully emit on transfer of ownership or drug or the store. Next i removed the isAdmin from addDrug() as we have a constructor. next I remodel the isUserAdmin() to actually check if the sender is same as the admin and reture a bool. On the reportDrug(), i left it like that as the main idea is to charge reporters to reduce the number of unnecessary report as u mentioned. Next changeDrugOwnership() now only change a ownership of a particular drug. While the new changeStoreOwnership() change ownership of a drug store. Next I updated the isSold behavior still keeping you idea on using bool. if set to true we can now have an emit of drugOutofStock().